requirements.txt
python-telegram-bot==20.8
aiohttp==3.9.1